server/API/requirements.txt
aniso8601==8.0.0
attrs==20.2.0
automium==0.2.6
automium-web==0.1.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dataclasses==0.6
efficientnet-pytorch==0.7.0
Flask==1.1.2
Flask-Cors==3.0.9
flask-restplus==0.13.0
future==0.18.2
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
legacy==0.1.6
lmdb==1.0.0
MarkupSafe==1.1.1
numpy==1.19.3
opencv-python==4.4.0.44
pandas==1.1.4
Pillow==8.0.1
protobuf==3.13.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
torch==1.7.0
torchaudio==0.7.0
torchtoolbox==0.1.5
torchvision==0.8.1
tqdm==4.51.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==0.16.1
gunicorn
